requirements.txt
aiohttp
yarl
pillow
telethon
cryptg
python-magic
rlottie-python
